+++ observer/ObserverPattern.ConsoleApp/WeatherData/WeatherDataState.cs
@@ -2,11 +2,11 @@ namespace ObserverPattern.Domain;
 
 public class WeatherDataState
 {
-  public float Temperature { get; }
-  public float Pressure { get; }
-  public float Humidity { get; }
+  public double Temperature { get; }
+  public double Pressure { get; }
+  public double Humidity { get; }
 
-  public WeatherDataState(float temperature, float pressure, float humidity)
+  public WeatherDataState(double temperature, double pressure, double humidity)
   {
     this.Temperature = temperature;
     this.Pressure = pressure;
